Added error handling in user interface.

+++ src/database/database.js
@@ -31,10 +31,10 @@ let dbPromise = null;
 const create = async () => {
   console.log('DatabaseService: Destroying old database..');
   // Only perform when performing E2E test
-  /* if (window.Cypress) {
+  if (window.Cypress) {
     await RxDB.removeDatabase('mailgenerator', 'idb');
-  } */
-  await RxDB.removeDatabase('mailgenerator', 'idb');
+  }
+  // await RxDB.removeDatabase('mailgenerator', 'idb');
   console.log('DatabaseService: Creating database..');
   const db = await RxDB.create({
     name: 'mailgenerator',
